chore: objects split into separate files
