+++ src/BoundType.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Thenativeweb\Eventsourcingdb;
+
+enum BoundType: string
+{
+    case INCLUSIVE = 'inclusive';
+    case EXCLUSIVE = 'exclusive';
+}
