+++ src/ObserveEventsOptions.php
@@ -6,31 +6,6 @@ namespace Thenativeweb\Eventsourcingdb;
 
 use JsonSerializable;
 
-enum ObserveIfEventIsMissing: string
-{
-    case READ_EVERYTHING = 'read-everything';
-    case WAIT_FOR_EVENT = 'wait-for-event';
-}
-
-final readonly class ObserveFromLatestEvent implements JsonSerializable
-{
-    public function __construct(
-        public string $subject,
-        public string $type,
-        public ObserveIfEventIsMissing $ifEventIsMissing,
-    ) {
-    }
-
-    public function jsonSerialize(): array
-    {
-        return [
-            'subject' => $this->subject,
-            'type' => $this->type,
-            'ifEventIsMissing' => $this->ifEventIsMissing->value,
-        ];
-    }
-}
-
 final readonly class ObserveEventsOptions implements JsonSerializable
 {
     public function __construct(
