+++ src/Order.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Thenativeweb\Eventsourcingdb;
+
+enum Order: string
+{
+    case CHRONOLOGICAL = 'chronological';
+    case ANTICHRONOLOGICAL = 'antichronological';
+}
